Extract prompt to its own file
