+++ next-app/app/api/chat/data.ts
@@ -1,409 +1,181 @@
-export const productCatalog = `Date (GMT),Job Link,Origin URL,Product Features Limit,Product Feature Sub Categories Limit,Reviews of Product Limit,Gender Cloths,Type of Clothes,Product Title,Extra Product Title Description,New or Used,price,Color,Available Color,Select Size,Sizes Available,Review Score
-2023-06-24 17:29:05,https://dashboard.browse.ai/teams/personal/robots/f701254f-ca74-4f67-b8cc-9f8b84ca4bae/history/644900f9-8182-48d4-adac-d41fc450db90,https://shop.lululemon.com/p/women-maintops/Swiftly-Tech-LS-2-Race/_/prod9820596?color=62255,4,4,5,Women's Clothes,Shirts,Swiftly Tech Long-Sleeve Shirt 2.0,Race Length,NEW,$78 CAD,Colour,"<div data-testid=""swatch"" aria-checked=""true"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""0"" aria-label=""Dark Lavender/Dark Lavender"" id=""swatch-0"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;tops-long-sleeve&quot;,&quot;color&quot;:&quot;62255&quot;,&quot;name&quot;:&quot;Swiftly Tech Long-Sleeve Shirt 2.0 *Race Length&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW3DOBS-062255&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi colorSwatchActive-3hOqb""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/62255?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/62255?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/62255?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/62255?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Dark Lavender/Dark Lavender"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/62255?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/62255?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Dark Lavender/Dark Lavender""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Poolside/Poolside"" id=""swatch-1"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;tops-long-sleeve&quot;,&quot;color&quot;:&quot;55316&quot;,&quot;name&quot;:&quot;Swiftly Tech Long-Sleeve Shirt 2.0 *Race Length&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW3DOBS-055316&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/55316?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/55316?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/55316?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/55316?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Poolside/Poolside"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/55316?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/55316?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Poolside/Poolside""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Cyan Blue/Cyan Blue"" id=""swatch-2"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;tops-long-sleeve&quot;,&quot;color&quot;:&quot;61906&quot;,&quot;name&quot;:&quot;Swiftly Tech Long-Sleeve Shirt 2.0 *Race Length&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW3DOBS-061906&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/61906?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/61906?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/61906?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/61906?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Cyan Blue/Cyan Blue"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/61906?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/61906?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Cyan Blue/Cyan Blue""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Kelly Green/Kelly Green"" id=""swatch-3"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;tops-long-sleeve&quot;,&quot;color&quot;:&quot;61924&quot;,&quot;name&quot;:&quot;Swiftly Tech Long-Sleeve Shirt 2.0 *Race Length&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW3HHIS-061924&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/61924?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/61924?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/61924?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/61924?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Kelly Green/Kelly Green"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/61924?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/61924?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Kelly Green/Kelly Green""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Slate/White"" id=""swatch-4""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/0572?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/0572?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/0572?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/0572?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Slate/White"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/0572?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/0572?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Slate/White""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Tetra Stripe Asphalt Grey/Black/Alpine White/Rhino Grey"" id=""swatch-5"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;tops-long-sleeve&quot;,&quot;color&quot;:&quot;50284&quot;,&quot;name&quot;:&quot;Swiftly Tech Long-Sleeve Shirt 2.0 *Race Length&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW3DPHS-050284&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/50284?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/50284?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/50284?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/50284?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Tetra Stripe Asphalt Grey/Black/Alpine White/Rhino Grey"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/50284?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/50284?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Tetra Stripe Asphalt Grey/Black/Alpine White/Rhino Grey""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Lilac Smoke/Lilac Smoke"" id=""swatch-6"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;tops-long-sleeve&quot;,&quot;color&quot;:&quot;61908&quot;,&quot;name&quot;:&quot;Swiftly Tech Long-Sleeve Shirt 2.0 *Race Length&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW3HHIS-061908&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/61908?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/61908?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/61908?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/61908?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Lilac Smoke/Lilac Smoke"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/61908?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/61908?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Lilac Smoke/Lilac Smoke""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Lip Gloss/Lip Gloss"" id=""swatch-7"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;tops-long-sleeve&quot;,&quot;color&quot;:&quot;28854&quot;,&quot;name&quot;:&quot;Swiftly Tech Long-Sleeve Shirt 2.0 *Race Length&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW3DOBS-028854&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/28854?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/28854?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/28854?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/28854?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Lip Gloss/Lip Gloss"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/28854?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/28854?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Lip Gloss/Lip Gloss""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Electric Lemon/Electric Lemon"" id=""swatch-8"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;tops-long-sleeve&quot;,&quot;color&quot;:&quot;53777&quot;,&quot;name&quot;:&quot;Swiftly Tech Long-Sleeve Shirt 2.0 *Race Length&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW3DOBS-053777&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/53777?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/53777?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/53777?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/53777?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Electric Lemon/Electric Lemon"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/53777?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/53777?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Electric Lemon/Electric Lemon""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Sonic Pink/Sonic Pink"" id=""swatch-9"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;tops-long-sleeve&quot;,&quot;color&quot;:&quot;35640&quot;,&quot;name&quot;:&quot;Swiftly Tech Long-Sleeve Shirt 2.0 *Race Length&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW3DOBS-035640&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/35640?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/35640?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/35640?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/35640?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Sonic Pink/Sonic Pink"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/35640?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/35640?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Sonic Pink/Sonic Pink""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Pink Peony/Velvet Dust"" id=""swatch-10"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;tops-long-sleeve&quot;,&quot;color&quot;:&quot;60535&quot;,&quot;name&quot;:&quot;Swiftly Tech Long-Sleeve Shirt 2.0 *Race Length&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW3DOBS-060535&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/60535?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/60535?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/60535?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/60535?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Pink Peony/Velvet Dust"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/60535?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/60535?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Pink Peony/Velvet Dust""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Dark Red/Dark Red"" id=""swatch-11""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/32007?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/32007?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/32007?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/32007?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Dark Red/Dark Red"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/32007?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/32007?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Dark Red/Dark Red""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Pink Peony/Pink Peony"" id=""swatch-12"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;tops-long-sleeve&quot;,&quot;color&quot;:&quot;57569&quot;,&quot;name&quot;:&quot;Swiftly Tech Long-Sleeve Shirt 2.0 *Race Length&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW3DOBS-057569&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/57569?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/57569?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/57569?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/57569?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Pink Peony/Pink Peony"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/57569?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/57569?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Pink Peony/Pink Peony""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Elixir/Elixir"" id=""swatch-13"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;tops-long-sleeve&quot;,&quot;color&quot;:&quot;38541&quot;,&quot;name&quot;:&quot;Swiftly Tech Long-Sleeve Shirt 2.0 *Race Length&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW3DOBS-038541&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/38541?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/38541?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/38541?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/38541?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Elixir/Elixir"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/38541?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/38541?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Elixir/Elixir""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Powder Blue/Powder Blue"" id=""swatch-14""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/59228?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/59228?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/59228?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/59228?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Powder Blue/Powder Blue"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/59228?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/59228?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Powder Blue/Powder Blue""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Water Drop/Water Drop"" id=""swatch-15"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;tops-long-sleeve&quot;,&quot;color&quot;:&quot;48666&quot;,&quot;name&quot;:&quot;Swiftly Tech Long-Sleeve Shirt 2.0 *Race Length&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW3DOBS-048666&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/48666?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/48666?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/48666?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/48666?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Water Drop/Water Drop"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/48666?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/48666?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Water Drop/Water Drop""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""White/White"" id=""swatch-16""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/12826?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/12826?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/12826?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/12826?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""White/White"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/12826?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/12826?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""White/White""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Black/Black"" id=""swatch-17"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;tops-long-sleeve&quot;,&quot;color&quot;:&quot;4780&quot;,&quot;name&quot;:&quot;Swiftly Tech Long-Sleeve Shirt 2.0 *Race Length&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW3DOBS-4780&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/4780?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/4780?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/4780?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/4780?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Black/Black"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/4780?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/4780?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Black/Black""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""True Navy/Iron Blue"" id=""swatch-18"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;tops-long-sleeve&quot;,&quot;color&quot;:&quot;46597&quot;,&quot;name&quot;:&quot;Swiftly Tech Long-Sleeve Shirt 2.0 *Race Length&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW3DOBS-046597&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/46597?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/46597?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/46597?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/46597?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""True Navy/Iron Blue"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/46597?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/46597?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""True Navy/Iron Blue""></picture></span></span></div>",Select Size,"0
-2
-4
-6
-8
-10
-12
-14
-16
-18
-20",4.4
-2023-06-24 17:29:09,https://dashboard.browse.ai/teams/personal/robots/f701254f-ca74-4f67-b8cc-9f8b84ca4bae/history/cfe4e90e-07b8-47c1-8431-e73bfab95904,https://shop.lululemon.com/p/women-tanks/Align-Ribbed-Tank/_/prod11250297?color=60714,4,4,5,Women's Clothes,Shirts,lululemon Align™ Ribbed Tank Top,,,$68 CAD,Colour,"<div data-testid=""swatch"" aria-checked=""true"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""0"" aria-label=""Dark Lavender"" id=""swatch-0""><span class=""colorSwatch-2yqfi colorSwatchActive-3hOqb""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/60714?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/60714?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/60714?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/60714?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Dark Lavender"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/60714?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/60714?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Dark Lavender""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Black"" id=""swatch-1"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;women-tanks&quot;,&quot;color&quot;:&quot;0001&quot;,&quot;name&quot;:&quot;lululemon Align™ Ribbed Tank Top&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW1DVLS-0001&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/0001?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/0001?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/0001?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/0001?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Black"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/0001?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/0001?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Black""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""White"" id=""swatch-2""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/0002?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/0002?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/0002?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/0002?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""White"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/0002?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/0002?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""White""></picture></span></span></div>",Select Size,"0
-2
-4
-6
-8
-10
-12
-14
-16
-18
-20",4.5
-2023-06-24 17:30:20,https://dashboard.browse.ai/teams/personal/robots/f701254f-ca74-4f67-b8cc-9f8b84ca4bae/history/8d594eff-f74c-4344-83bd-f9af727dab05,https://shop.lululemon.com/p/women-tanks/Sculpt-Tank-Cropped/_/prod9961111?color=28936,4,4,5,Women's Clothes,Shirts,Sculpt Cropped Tank Top,,,$58 CAD,Colour,"<div data-testid=""swatch"" aria-checked=""true"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""0"" aria-label=""Lip Gloss"" id=""swatch-0"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;women-tanks&quot;,&quot;color&quot;:&quot;28936&quot;,&quot;name&quot;:&quot;Sculpt Cropped Tank Top&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW1DZSS-028936&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi colorSwatchActive-3hOqb""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/28936?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/28936?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/28936?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/28936?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Lip Gloss"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/28936?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/28936?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Lip Gloss""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Kelly Green"" id=""swatch-1"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;women-tanks&quot;,&quot;color&quot;:&quot;60997&quot;,&quot;name&quot;:&quot;Sculpt Cropped Tank Top&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW1DZSS-060997&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/60997?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/60997?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/60997?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/60997?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Kelly Green"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/60997?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/60997?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Kelly Green""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Bone"" id=""swatch-2""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/27597?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/27597?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/27597?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/27597?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Bone"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/27597?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/27597?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Bone""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Mango Dream"" id=""swatch-3"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;women-tanks&quot;,&quot;color&quot;:&quot;60715&quot;,&quot;name&quot;:&quot;Sculpt Cropped Tank Top&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW1DZSS-060715&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/60715?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/60715?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/60715?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/60715?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Mango Dream"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/60715?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/60715?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Mango Dream""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Black"" id=""swatch-4"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;women-tanks&quot;,&quot;color&quot;:&quot;0001&quot;,&quot;name&quot;:&quot;Sculpt Cropped Tank Top&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW1DZSS-0001&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/0001?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/0001?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/0001?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/0001?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Black"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/0001?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/0001?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Black""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Paint Drift Multi"" id=""swatch-5"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;women-tanks&quot;,&quot;color&quot;:&quot;61871&quot;,&quot;name&quot;:&quot;Sculpt Cropped Tank Top&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW1EN2S-061871&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/61871?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/61871?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/61871?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/61871?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Paint Drift Multi"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/61871?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/61871?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Paint Drift Multi""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Lilac Smoke"" id=""swatch-6"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;women-tanks&quot;,&quot;color&quot;:&quot;55325&quot;,&quot;name&quot;:&quot;Sculpt Cropped Tank Top&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW1DZSS-055325&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/55325?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/55325?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/55325?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/55325?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Lilac Smoke"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/55325?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/55325?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Lilac Smoke""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Cyan Blue"" id=""swatch-7"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;women-tanks&quot;,&quot;color&quot;:&quot;49226&quot;,&quot;name&quot;:&quot;Sculpt Cropped Tank Top&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW1DZSS-049226&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/49226?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/49226?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/49226?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/49226?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Cyan Blue"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/49226?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/49226?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Cyan Blue""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Sonic Pink"" id=""swatch-8"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;women-tanks&quot;,&quot;color&quot;:&quot;36161&quot;,&quot;name&quot;:&quot;Sculpt Cropped Tank Top&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW1DZSS-036161&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/36161?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/36161?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/36161?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/36161?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Sonic Pink"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/36161?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/36161?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Sonic Pink""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""White"" id=""swatch-9"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;women-tanks&quot;,&quot;color&quot;:&quot;0002&quot;,&quot;name&quot;:&quot;Sculpt Cropped Tank Top&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW1DZSS-0002&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/0002?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/0002?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/0002?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/0002?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""White"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/0002?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/0002?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""White""></picture></span></span></div>",Select Size,"0
-2
-4
-6
-8
-10
-12
-14
-16
-18
- (not available)
-20
- (not available)",4.4
-2023-06-24 17:31:45,https://dashboard.browse.ai/teams/personal/robots/f701254f-ca74-4f67-b8cc-9f8b84ca4bae/history/485eb441-37e5-4dc0-bf77-422c72645442,https://shop.lululemon.com/p/women-sports-bras/Align-V-Neck-Bra-CD-Cup/_/prod11440041?color=28936,4,4,5,Women's Clothes,Sports Bras,lululemon Align™ V-Neck Bra,"Light Support, C/D Cup",NEW,$64 CAD,Colour,"<div data-testid=""swatch"" aria-checked=""true"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""0"" aria-label=""Lip Gloss"" id=""swatch-0"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;women-sports-bras&quot;,&quot;color&quot;:&quot;28936&quot;,&quot;name&quot;:&quot;lululemon Align™ V-Neck Bra *Light Support, C/D Cup&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW2DTJS-028936&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi colorSwatchActive-3hOqb""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/28936?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/28936?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/28936?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/28936?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Lip Gloss"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/28936?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/28936?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Lip Gloss""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Dark Lavender"" id=""swatch-1"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;women-sports-bras&quot;,&quot;color&quot;:&quot;60714&quot;,&quot;name&quot;:&quot;lululemon Align™ V-Neck Bra *Light Support, C/D Cup&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW2DTJS-060714&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/60714?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/60714?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/60714?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/60714?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Dark Lavender"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/60714?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/60714?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Dark Lavender""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Lilac Smoke"" id=""swatch-2"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;women-sports-bras&quot;,&quot;color&quot;:&quot;55325&quot;,&quot;name&quot;:&quot;lululemon Align™ V-Neck Bra *Light Support, C/D Cup&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW2DTJS-055325&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/55325?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/55325?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/55325?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/55325?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Lilac Smoke"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/55325?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/55325?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Lilac Smoke""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Twilight Rose"" id=""swatch-3""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/55138?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/55138?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/55138?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/55138?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Twilight Rose"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/55138?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/55138?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Twilight Rose""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Kelly Green"" id=""swatch-4"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;women-sports-bras&quot;,&quot;color&quot;:&quot;60997&quot;,&quot;name&quot;:&quot;lululemon Align™ V-Neck Bra *Light Support, C/D Cup&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW2DTJS-060997&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/60997?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/60997?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/60997?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/60997?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Kelly Green"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/60997?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/60997?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Kelly Green""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Cheetah Camo Antique White Multi"" id=""swatch-5"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;women-sports-bras&quot;,&quot;color&quot;:&quot;56012&quot;,&quot;name&quot;:&quot;lululemon Align™ V-Neck Bra *Light Support, C/D Cup&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW2DX7S-056012&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/56012?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/56012?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/56012?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/56012?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Cheetah Camo Antique White Multi"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/56012?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/56012?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Cheetah Camo Antique White Multi""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""White"" id=""swatch-6"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;women-sports-bras&quot;,&quot;color&quot;:&quot;0002&quot;,&quot;name&quot;:&quot;lululemon Align™ V-Neck Bra *Light Support, C/D Cup&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW2DTJS-0002&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/0002?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/0002?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/0002?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/0002?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""White"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/0002?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/0002?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""White""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Utility Blue"" id=""swatch-7"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;women-sports-bras&quot;,&quot;color&quot;:&quot;36744&quot;,&quot;name&quot;:&quot;lululemon Align™ V-Neck Bra *Light Support, C/D Cup&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW2DPOS-036744&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/36744?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/36744?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/36744?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/36744?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Utility Blue"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/36744?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/36744?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Utility Blue""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Poolside"" id=""swatch-8""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/54319?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/54319?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/54319?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/54319?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Poolside"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/54319?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/54319?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Poolside""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Pink Peony"" id=""swatch-9"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;women-sports-bras&quot;,&quot;color&quot;:&quot;56496&quot;,&quot;name&quot;:&quot;lululemon Align™ V-Neck Bra *Light Support, C/D Cup&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW2DTJS-056496&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/56496?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/56496?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/56496?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/56496?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Pink Peony"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/56496?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/56496?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Pink Peony""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Mango Dream"" id=""swatch-10"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;women-sports-bras&quot;,&quot;color&quot;:&quot;60715&quot;,&quot;name&quot;:&quot;lululemon Align™ V-Neck Bra *Light Support, C/D Cup&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW2DTJS-060715&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/60715?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/60715?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/60715?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/60715?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Mango Dream"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/60715?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/60715?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Mango Dream""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Black"" id=""swatch-11"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;women-sports-bras&quot;,&quot;color&quot;:&quot;0001&quot;,&quot;name&quot;:&quot;lululemon Align™ V-Neck Bra *Light Support, C/D Cup&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW2DPOS-0001&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/0001?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/0001?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/0001?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/0001?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Black"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/0001?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/0001?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Black""></picture></span></span></div>",Select Size,"2
-4
-6
-8
-10
-12
-14",4
-2023-06-24 17:32:22,https://dashboard.browse.ai/teams/personal/robots/f701254f-ca74-4f67-b8cc-9f8b84ca4bae/history/ef4b6871-f023-4e41-8a57-af19d6b42c4e,https://shop.lululemon.com/p/women-pants/Nulux-5-Pocket-HR-Run-Tight-25-Ref/_/prod11380248?color=60997,4,4,5,Women's Clothes,Pants,Fast and Free High-Rise Tight 25” Pockets,Updated,,$138 CAD,Colour,"<div data-testid=""swatch"" aria-checked=""true"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""0"" aria-label=""Kelly Green"" id=""swatch-0""><span class=""colorSwatch-2yqfi colorSwatchActive-3hOqb""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/60997?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/60997?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/60997?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/60997?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Kelly Green"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/60997?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/60997?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Kelly Green""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Paint Drift Multi"" id=""swatch-1"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;women-pants&quot;,&quot;color&quot;:&quot;61871&quot;,&quot;name&quot;:&quot;Fast and Free High-Rise Tight 25” Pockets *Updated&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW5FODS-061871&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/61871?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/61871?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/61871?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/61871?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Paint Drift Multi"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/61871?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/61871?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Paint Drift Multi""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Wild Indigo"" id=""swatch-2""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/58292?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/58292?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/58292?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/58292?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Wild Indigo"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/58292?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/58292?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Wild Indigo""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""True Navy"" id=""swatch-3"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;women-pants&quot;,&quot;color&quot;:&quot;31382&quot;,&quot;name&quot;:&quot;Fast and Free High-Rise Tight 25” Pockets *Updated&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW5FARS-031382&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/31382?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/31382?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/31382?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/31382?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""True Navy"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/31382?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/31382?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""True Navy""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Black"" id=""swatch-4"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;women-pants&quot;,&quot;color&quot;:&quot;0001&quot;,&quot;name&quot;:&quot;Fast and Free High-Rise Tight 25” Pockets *Updated&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW5FARS-0001&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/0001?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/0001?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/0001?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/0001?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Black"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/0001?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/0001?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Black""></picture></span></span></div>",Select Size,"0
-2
-4
-6
-8
-10
-12
-14
-16
-18
- (not available)
-20
- (not available)",2.9
-2023-06-24 17:32:26,https://dashboard.browse.ai/teams/personal/robots/f701254f-ca74-4f67-b8cc-9f8b84ca4bae/history/ce03e395-e5a8-4fa7-877f-e2a0313adf0d,https://shop.lululemon.com/p/women-shorts/Align-Short-6-Pocket/_/prod10520166?color=28936,4,4,5,Women's Clothes,Shorts,"lululemon Align™ High-Rise Short with Pockets 6""",Online Only,,$74 CAD,Colour,"<div data-testid=""swatch"" aria-checked=""true"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""0"" aria-label=""Lip Gloss"" id=""swatch-0"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;women-shorts&quot;,&quot;color&quot;:&quot;28936&quot;,&quot;name&quot;:&quot;lululemon Align™ High-Rise Short with Pockets 6\&quot; *Online Only&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW7CRQS-028936&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi colorSwatchActive-3hOqb""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/28936?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/28936?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/28936?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/28936?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Lip Gloss"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/28936?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/28936?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Lip Gloss""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Cyan Blue"" id=""swatch-1"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;women-shorts&quot;,&quot;color&quot;:&quot;49226&quot;,&quot;name&quot;:&quot;lululemon Align™ High-Rise Short with Pockets 6\&quot; *Online Only&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW7CRQS-049226&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/49226?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/49226?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/49226?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/49226?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Cyan Blue"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/49226?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/49226?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Cyan Blue""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Java"" id=""swatch-2"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;women-shorts&quot;,&quot;color&quot;:&quot;52868&quot;,&quot;name&quot;:&quot;lululemon Align™ High-Rise Short with Pockets 6\&quot; *Online Only&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW7CRPS-052868&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/52868?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/52868?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/52868?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/52868?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Java"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/52868?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/52868?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Java""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Dark Lavender"" id=""swatch-3"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;women-shorts&quot;,&quot;color&quot;:&quot;60714&quot;,&quot;name&quot;:&quot;lululemon Align™ High-Rise Short with Pockets 6\&quot; *Online Only&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW7CTHS-060714&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/60714?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/60714?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/60714?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/60714?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Dark Lavender"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/60714?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/60714?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Dark Lavender""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Black"" id=""swatch-4""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/0001?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/0001?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/0001?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/0001?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Black"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/0001?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/0001?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Black""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Utility Blue"" id=""swatch-5"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;women-shorts&quot;,&quot;color&quot;:&quot;36744&quot;,&quot;name&quot;:&quot;lululemon Align™ High-Rise Short with Pockets 6\&quot; *Online Only&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW7CRPS-036744&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/36744?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/36744?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/36744?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/36744?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Utility Blue"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/36744?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/36744?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Utility Blue""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""True Navy"" id=""swatch-6"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;women-shorts&quot;,&quot;color&quot;:&quot;31382&quot;,&quot;name&quot;:&quot;lululemon Align™ High-Rise Short with Pockets 6\&quot; *Online Only&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW7BUMS-031382&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/31382?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/31382?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/31382?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/31382?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""True Navy"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/31382?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/31382?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""True Navy""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Heritage 365 Camo Deep Coal Multi"" id=""swatch-7"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;women-shorts&quot;,&quot;color&quot;:&quot;47184&quot;,&quot;name&quot;:&quot;lululemon Align™ High-Rise Short with Pockets 6\&quot; *Online Only&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW7CBKS-047184&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/47184?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/47184?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/47184?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/47184?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Heritage 365 Camo Deep Coal Multi"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/47184?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/47184?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Heritage 365 Camo Deep Coal Multi""></picture></span></span></div>",Select Size,"0
-2
-4
-6
-8
-10
-12
-14",4.5
-2023-06-24 17:33:04,https://dashboard.browse.ai/teams/personal/robots/f701254f-ca74-4f67-b8cc-9f8b84ca4bae/history/3d09622c-e532-421a-a541-b01ea07bc350,https://shop.lululemon.com/p/women-sports-bras/Like-a-Cloud-LL-Ribbed-D/_/prod11021146?color=60997,4,4,5,Women's Clothes,Sports Bras,Like a Cloud Longline Ribbed Bra,"Light Support, D/DD Cups",,$74 CAD,Colour,"<div data-testid=""swatch"" aria-checked=""true"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""0"" aria-label=""Kelly Green"" id=""swatch-0"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;women&quot;,&quot;color&quot;:&quot;60997&quot;,&quot;name&quot;:&quot;Like a Cloud Longline Ribbed Bra *Light Support, D/DD Cups&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW2DJES-060997&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi colorSwatchActive-3hOqb""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/60997?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/60997?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/60997?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/60997?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Kelly Green"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/60997?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/60997?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Kelly Green""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Pink Peony"" id=""swatch-1"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;women&quot;,&quot;color&quot;:&quot;56496&quot;,&quot;name&quot;:&quot;Like a Cloud Longline Ribbed Bra *Light Support, D/DD Cups&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW2DJES-056496&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/56496?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/56496?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/56496?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/56496?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Pink Peony"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/56496?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/56496?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Pink Peony""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Dark Lavender"" id=""swatch-2"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;women&quot;,&quot;color&quot;:&quot;60714&quot;,&quot;name&quot;:&quot;Like a Cloud Longline Ribbed Bra *Light Support, D/DD Cups&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW2DJES-060714&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/60714?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/60714?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/60714?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/60714?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Dark Lavender"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/60714?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/60714?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Dark Lavender""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""White"" id=""swatch-3"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;women&quot;,&quot;color&quot;:&quot;0002&quot;,&quot;name&quot;:&quot;Like a Cloud Longline Ribbed Bra *Light Support, D/DD Cups&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW2DJES-0002&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/0002?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/0002?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/0002?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/0002?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""White"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/0002?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/0002?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""White""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Black"" id=""swatch-4"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;women&quot;,&quot;color&quot;:&quot;0001&quot;,&quot;name&quot;:&quot;Like a Cloud Longline Ribbed Bra *Light Support, D/DD Cups&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW2DJES-0001&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/0001?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/0001?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/0001?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/0001?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Black"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/0001?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/0001?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Black""></picture></span></span></div>",Select Size,"2
-4
-6
-8
-10
-12
-14",4.6
-2023-06-24 17:33:07,https://dashboard.browse.ai/teams/personal/robots/f701254f-ca74-4f67-b8cc-9f8b84ca4bae/history/e0eba237-0a92-47ef-b871-f68accde7228,https://shop.lululemon.com/p/skirts-and-dresses-skirts/Pace-Rival-Skirt-Tall/_/prod3770001?color=60997,4,4,5,Women's Clothes,Skirts,Pace Rival Mid-Rise Skirt,Long,NEW,$78 CAD,Colour,"<div data-testid=""swatch"" aria-checked=""true"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""0"" aria-label=""Kelly Green"" id=""swatch-0"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;skirts-and-dresses-skirts&quot;,&quot;color&quot;:&quot;60997&quot;,&quot;name&quot;:&quot;Pace Rival Mid-Rise Skirt *Long&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW8AC7T-060997&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi colorSwatchActive-3hOqb""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/60997?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/60997?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/60997?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/60997?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Kelly Green"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/60997?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/60997?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Kelly Green""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Kohlrabi Green"" id=""swatch-1"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;skirts-and-dresses-skirts&quot;,&quot;color&quot;:&quot;60300&quot;,&quot;name&quot;:&quot;Pace Rival Mid-Rise Skirt *Long&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW8AC7T-060300&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/60300?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/60300?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/60300?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/60300?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Kohlrabi Green"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/60300?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/60300?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Kohlrabi Green""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Lip Gloss"" id=""swatch-2"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;skirts-and-dresses-skirts&quot;,&quot;color&quot;:&quot;28936&quot;,&quot;name&quot;:&quot;Pace Rival Mid-Rise Skirt *Long&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW8AC7T-028936&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/28936?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/28936?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/28936?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/28936?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Lip Gloss"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/28936?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/28936?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Lip Gloss""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Poolside"" id=""swatch-3"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;skirts-and-dresses-skirts&quot;,&quot;color&quot;:&quot;54319&quot;,&quot;name&quot;:&quot;Pace Rival Mid-Rise Skirt *Long&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW8AACT-054319&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/54319?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/54319?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/54319?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/54319?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Poolside"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/54319?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/54319?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Poolside""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Mango Dream"" id=""swatch-4"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;skirts-and-dresses-skirts&quot;,&quot;color&quot;:&quot;60715&quot;,&quot;name&quot;:&quot;Pace Rival Mid-Rise Skirt *Long&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW8AC7T-060715&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/60715?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/60715?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/60715?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/60715?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Mango Dream"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/60715?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/60715?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Mango Dream""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Hot Heat"" id=""swatch-5"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;skirts-and-dresses-skirts&quot;,&quot;color&quot;:&quot;34668&quot;,&quot;name&quot;:&quot;Pace Rival Mid-Rise Skirt *Long&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW8AC7T-034668&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/34668?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/34668?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/34668?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/34668?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Hot Heat"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/34668?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/34668?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Hot Heat""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Sonic Pink"" id=""swatch-6""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/36161?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/36161?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/36161?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/36161?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Sonic Pink"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/36161?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/36161?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Sonic Pink""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""True Navy"" id=""swatch-7""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/31382?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/31382?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/31382?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/31382?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""True Navy"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/31382?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/31382?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""True Navy""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""White"" id=""swatch-8"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;skirts-and-dresses-skirts&quot;,&quot;color&quot;:&quot;0002&quot;,&quot;name&quot;:&quot;Pace Rival Mid-Rise Skirt *Long&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW8A85T-0002&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/0002?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/0002?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/0002?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/0002?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""White"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/0002?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/0002?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""White""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Pastel Blue"" id=""swatch-9"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;skirts-and-dresses-skirts&quot;,&quot;color&quot;:&quot;52291&quot;,&quot;name&quot;:&quot;Pace Rival Mid-Rise Skirt *Long&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW8AADT-052291&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/52291?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/52291?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/52291?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/52291?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Pastel Blue"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/52291?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/52291?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Pastel Blue""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Black"" id=""swatch-10"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;skirts-and-dresses-skirts&quot;,&quot;color&quot;:&quot;0001&quot;,&quot;name&quot;:&quot;Pace Rival Mid-Rise Skirt *Long&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW8AACT-0001&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/0001?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/0001?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/0001?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/0001?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Black"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/0001?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/0001?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Black""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Dark Red"" id=""swatch-11"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;skirts-and-dresses-skirts&quot;,&quot;color&quot;:&quot;28948&quot;,&quot;name&quot;:&quot;Pace Rival Mid-Rise Skirt *Long&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW8AACT-028948&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/28948?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/28948?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/28948?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/28948?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Dark Red"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/28948?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/28948?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Dark Red""></picture></span></span></div>",Select Size,"0
-2
-4
-6
-8
-10
-12
-14
-16
- (not available)
-18
- (not available)
-20
- (not available)",4.4
-2023-06-24 17:33:39,https://dashboard.browse.ai/teams/personal/robots/f701254f-ca74-4f67-b8cc-9f8b84ca4bae/history/da9153ca-8b44-40b1-8ae0-ee241a77cdc3,https://shop.lululemon.com/p/women-tanks/Sleeveless-Polo-Shirt/_/prod10930138?color=60997,4,4,5,Women's Clothes,Shirts,Quick-Dry Sleeveless Polo Shirt,,,$68 CAD,Colour,"<div data-testid=""swatch"" aria-checked=""true"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""0"" aria-label=""Kelly Green"" id=""swatch-0"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;women&quot;,&quot;color&quot;:&quot;60997&quot;,&quot;name&quot;:&quot;Quick-Dry Sleeveless Polo Shirt&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW1DWKS-060997&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi colorSwatchActive-3hOqb""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/60997?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/60997?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/60997?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/60997?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Kelly Green"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/60997?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/60997?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Kelly Green""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""True Navy"" id=""swatch-1""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/31382?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/31382?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/31382?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/31382?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""True Navy"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/31382?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/31382?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""True Navy""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Terrazzo Marble Silver Ice Multi"" id=""swatch-2"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;women&quot;,&quot;color&quot;:&quot;61843&quot;,&quot;name&quot;:&quot;Quick-Dry Sleeveless Polo Shirt&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW1EBQS-061843&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/61843?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/61843?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/61843?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/61843?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Terrazzo Marble Silver Ice Multi"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/61843?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/61843?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Terrazzo Marble Silver Ice Multi""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""White"" id=""swatch-3"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;women&quot;,&quot;color&quot;:&quot;0002&quot;,&quot;name&quot;:&quot;Quick-Dry Sleeveless Polo Shirt&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW1DWKS-0002&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/0002?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/0002?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/0002?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/0002?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""White"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/0002?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/0002?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""White""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Lip Gloss"" id=""swatch-4"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;women&quot;,&quot;color&quot;:&quot;28936&quot;,&quot;name&quot;:&quot;Quick-Dry Sleeveless Polo Shirt&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW1DWKS-028936&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/28936?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/28936?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/28936?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/28936?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Lip Gloss"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/28936?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/28936?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Lip Gloss""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Summer Glow"" id=""swatch-5""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/60998?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/60998?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/60998?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/60998?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Summer Glow"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/60998?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/60998?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Summer Glow""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Black"" id=""swatch-6"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;women&quot;,&quot;color&quot;:&quot;0001&quot;,&quot;name&quot;:&quot;Quick-Dry Sleeveless Polo Shirt&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW1DWKS-0001&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/0001?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/0001?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/0001?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/0001?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Black"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/0001?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/0001?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Black""></picture></span></span></div>",Select Size,"0
-2
-4
-6
-8
-10
-12
- (not available)
-14
- (not available)
-16
- (not available)
-18
- (not available)
-20
- (not available)",4
-2023-06-24 17:33:42,https://dashboard.browse.ai/teams/personal/robots/f701254f-ca74-4f67-b8cc-9f8b84ca4bae/history/44dd8bcd-7c7f-42ce-9e56-3ab262761d8b,https://shop.lululemon.com/p/jackets-and-hoodies-jackets/Hooded-Define-Jacket-Nulu-2/_/prod11500123?color=61860,4,4,5,Women's Clothes,Coats & Jackets,Hooded Define Jacket,Nulu,NEW,$138 CAD,Colour,"<div data-testid=""swatch"" aria-checked=""true"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""0"" aria-label=""Meadowsweet Pink"" id=""swatch-0""><span class=""colorSwatch-2yqfi colorSwatchActive-3hOqb""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/61860?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/61860?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/61860?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/61860?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Meadowsweet Pink"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/61860?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/61860?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Meadowsweet Pink""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Black"" id=""swatch-1"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;whats-new&quot;,&quot;color&quot;:&quot;0001&quot;,&quot;name&quot;:&quot;Hooded Define Jacket *Nulu&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW3HAWS-0001&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/0001?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/0001?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/0001?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/0001?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Black"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/0001?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/0001?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Black""></picture></span></span></div>",Select Size,"0
-2
-4
-6
-8
-10
-12
-14",4.8
-2023-06-24 17:34:21,https://dashboard.browse.ai/teams/personal/robots/f701254f-ca74-4f67-b8cc-9f8b84ca4bae/history/c8d76950-33ea-4b09-8ea9-0bfaabdedcdb,https://shop.lululemon.com/p/women-sports-bras/Flow-Y-Bra-Nulu-Long-Line/_/prod9270365?color=56852,4,4,5,Women's Clothes,Sports Bras,Flow Y Nulu Longline Bra,"Light Support, A–C Cups",,$64 CAD,Colour,"<div data-testid=""swatch"" aria-checked=""true"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""0"" aria-label=""Flush Pink"" id=""swatch-0""><span class=""colorSwatch-2yqfi colorSwatchActive-3hOqb""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/56852?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/56852?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/56852?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/56852?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Flush Pink"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/56852?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/56852?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Flush Pink""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Kelly Green"" id=""swatch-1"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;women-sports-bras&quot;,&quot;color&quot;:&quot;60997&quot;,&quot;name&quot;:&quot;Flow Y Nulu Longline Bra *Light Support, A–C Cups *Online Only&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW2CZ7S-060997&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/60997?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/60997?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/60997?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/60997?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Kelly Green"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/60997?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/60997?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Kelly Green""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Poolside"" id=""swatch-2"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;women-sports-bras&quot;,&quot;color&quot;:&quot;54319&quot;,&quot;name&quot;:&quot;Flow Y Nulu Longline Bra *Light Support, A–C Cups *Online Only&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW2DWXS-054319&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/54319?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/54319?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/54319?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/54319?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Poolside"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/54319?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/54319?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Poolside""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Lip Gloss"" id=""swatch-3"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;women-sports-bras&quot;,&quot;color&quot;:&quot;28936&quot;,&quot;name&quot;:&quot;Flow Y Nulu Longline Bra *Light Support, A–C Cups *Online Only&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW2CZ7S-028936&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/28936?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/28936?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/28936?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/28936?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Lip Gloss"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/28936?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/28936?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Lip Gloss""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Lilac Smoke"" id=""swatch-4""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/55325?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/55325?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/55325?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/55325?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Lilac Smoke"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/55325?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/55325?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Lilac Smoke""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Black"" id=""swatch-5"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;women-sports-bras&quot;,&quot;color&quot;:&quot;0001&quot;,&quot;name&quot;:&quot;Flow Y Nulu Longline Bra *Light Support, A–C Cups *Online Only&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW2CZ6S-0001&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/0001?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/0001?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/0001?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/0001?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Black"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/0001?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/0001?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Black""></picture></span></span></div>",Select Size,"2
-4
-6
-8
-10
-12
-14",3.7
-2023-06-24 17:34:31,https://dashboard.browse.ai/teams/personal/robots/f701254f-ca74-4f67-b8cc-9f8b84ca4bae/history/316c3a58-eb51-4e36-b1d8-40b508fcaddf,https://shop.lululemon.com/p/womens-sweatpants/Studio-Crop-III-Unlined/_/prod2040016?color=56496,4,4,5,Women's Clothes,Pants,Dance Studio Mid-Rise Cropped Pant,,,$88 CAD,Colour,"<div data-testid=""swatch"" aria-checked=""true"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""0"" aria-label=""Pink Peony"" id=""swatch-0"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;women-crops&quot;,&quot;color&quot;:&quot;56496&quot;,&quot;name&quot;:&quot;Dance Studio Mid-Rise Cropped Pant&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW6CLWS-056496&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi colorSwatchActive-3hOqb""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/56496?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/56496?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/56496?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/56496?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Pink Peony"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/56496?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/56496?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Pink Peony""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Bone"" id=""swatch-1"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;women-crops&quot;,&quot;color&quot;:&quot;27597&quot;,&quot;name&quot;:&quot;Dance Studio Mid-Rise Cropped Pant&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW6CLWS-027597&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/27597?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/27597?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/27597?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/27597?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Bone"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/27597?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/27597?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Bone""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Utility Blue"" id=""swatch-2"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;women-crops&quot;,&quot;color&quot;:&quot;36744&quot;,&quot;name&quot;:&quot;Dance Studio Mid-Rise Cropped Pant&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW6CDQS-036744&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/36744?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/36744?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/36744?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/36744?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Utility Blue"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/36744?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/36744?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Utility Blue""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""Black"" id=""swatch-3"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;women-crops&quot;,&quot;color&quot;:&quot;0001&quot;,&quot;name&quot;:&quot;Dance Studio Mid-Rise Cropped Pant&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW6CDQS-0001&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/0001?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/0001?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/0001?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/0001?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""Black"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/0001?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/0001?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""Black""></picture></span></span></div><div data-testid=""swatch"" aria-checked=""false"" class=""buttonTile-2gzKp OneLinkNoTx available select-swatches_swatchButtonTile__jmyaF"" role=""radio"" tabindex=""-1"" aria-label=""True Navy"" id=""swatch-4"" data-lulu-attributes=""{&quot;type&quot;:&quot;product&quot;,&quot;product&quot;:{&quot;categoryUnifiedID&quot;:&quot;women-crops&quot;,&quot;color&quot;:&quot;31382&quot;,&quot;name&quot;:&quot;Dance Studio Mid-Rise Cropped Pant&quot;,&quot;price&quot;:&quot;&quot;,&quot;size&quot;:&quot;&quot;,&quot;styleID&quot;:&quot;LW6CDQS-031382&quot;},&quot;version&quot;:2}"" data-lulu-track=""pdp-miw-product-link""><span class=""colorSwatch-2yqfi""><span class=""colorSwatchImgWrapper-1k4IP""><picture class=""colorSwatchPicture-KEE-O""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/31382?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/31382?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=webp&amp;qlt=90,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/webp""><source sizes=""34px"" srcset=""https://images.lululemon.com/is/image/lululemon/31382?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/31382?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" type=""image/jpeg""><img alt=""True Navy"" class=""colorSwatchImg-2mYJB"" sizes=""34px"" src=""data:,"" srcset=""https://images.lululemon.com/is/image/lululemon/31382?wid=34&amp;hei=34&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 34w,
-https://images.lululemon.com/is/image/lululemon/31382?wid=68&amp;hei=68&amp;fit=crop,1&amp;op_usm=0.8,1,10,0&amp;fmt=jpeg&amp;qlt=75,1&amp;op_sharpen=0&amp;resMode=sharp2&amp;iccEmbed=0&amp;printRes=72 68w"" title=""True Navy""></picture></span></span></div>",Select Size,"0
-2
-4
-6
-8
-10
-12
- (not available)
-14
- (not available)",4.3
-`;
+export function getProductCatalog() {
+  const products = [
+    {
+      "Origin URL":
+        "https://shop.lululemon.com/p/women-maintops/Swiftly-Tech-LS-2-Race/_/prod9820596?color=62255",
+      "Product Features Limit": 4,
+      "Product Feature Sub Categories Limit": 4,
+      "Reviews of Product Limit": 5,
+      "Gender Cloths": "Women's Clothes",
+      "Type of Clothes": "Shirts",
+      "Product Title": "Swiftly Tech Long-Sleeve Shirt 2.0",
+      "Extra Product Title Description": "Race Length",
+      "New or Used": "NEW",
+      price: "$78 CAD",
+      "Review Score": 4.4,
+    },
+    {
+      "Origin URL":
+        "https://shop.lululemon.com/p/women-tanks/Align-Ribbed-Tank/_/prod11250297?color=60714",
+      "Product Features Limit": 4,
+      "Product Feature Sub Categories Limit": 4,
+      "Reviews of Product Limit": 5,
+      "Gender Cloths": "Women's Clothes",
+      "Type of Clothes": "Shirts",
+      "Product Title": "lululemon Align™ Ribbed Tank Top",
+      "Extra Product Title Description": "",
+      "New or Used": "",
+      price: "$68 CAD",
+      "Review Score": 4.5,
+    },
+    {
+      "Origin URL":
+        "https://shop.lululemon.com/p/women-tanks/Sculpt-Tank-Cropped/_/prod9961111?color=28936",
+      "Product Features Limit": 4,
+      "Product Feature Sub Categories Limit": 4,
+      "Reviews of Product Limit": 5,
+      "Gender Cloths": "Women's Clothes",
+      "Type of Clothes": "Shirts",
+      "Product Title": "Sculpt Cropped Tank Top",
+      "Extra Product Title Description": "",
+      "New or Used": "",
+      price: "$58 CAD",
+      "Review Score": 4.4,
+    },
+    {
+      "Origin URL":
+        "https://shop.lululemon.com/p/women-sports-bras/Align-V-Neck-Bra-CD-Cup/_/prod11440041?color=28936",
+      "Product Features Limit": 4,
+      "Product Feature Sub Categories Limit": 4,
+      "Reviews of Product Limit": 5,
+      "Gender Cloths": "Women's Clothes",
+      "Type of Clothes": "Sports Bras",
+      "Product Title": "lululemon Align™ V-Neck Bra",
+      "Extra Product Title Description": "Light Support, C/D Cup",
+      "New or Used": "NEW",
+      price: "$64 CAD",
+      "Review Score": 4,
+    },
+    {
+      "Origin URL":
+        "https://shop.lululemon.com/p/women-pants/Nulux-5-Pocket-HR-Run-Tight-25-Ref/_/prod11380248?color=60997",
+      "Product Features Limit": 4,
+      "Product Feature Sub Categories Limit": 4,
+      "Reviews of Product Limit": 5,
+      "Gender Cloths": "Women's Clothes",
+      "Type of Clothes": "Pants",
+      "Product Title": "Fast and Free High-Rise Tight 25” Pockets",
+      "Extra Product Title Description": "Updated",
+      "New or Used": "",
+      price: "$138 CAD",
+      "Review Score": 2.9,
+    },
+    {
+      "Origin URL":
+        "https://shop.lululemon.com/p/women-shorts/Align-Short-6-Pocket/_/prod10520166?color=28936",
+      "Product Features Limit": 4,
+      "Product Feature Sub Categories Limit": 4,
+      "Reviews of Product Limit": 5,
+      "Gender Cloths": "Women's Clothes",
+      "Type of Clothes": "Shorts",
+      "Product Title": "lululemon Align™ High-Rise Short with Pockets 6",
+      "Extra Product Title Description": "Online Only",
+      "New or Used": "",
+      price: "$74 CAD",
+      "Review Score": 4.5,
+    },
+    {
+      "Origin URL":
+        "https://shop.lululemon.com/p/women-sports-bras/Like-a-Cloud-LL-Ribbed-D/_/prod11021146?color=60997",
+      "Product Features Limit": 4,
+      "Product Feature Sub Categories Limit": 4,
+      "Reviews of Product Limit": 5,
+      "Gender Cloths": "Women's Clothes",
+      "Type of Clothes": "Sports Bras",
+      "Product Title": "Like a Cloud Longline Ribbed Bra",
+      "Extra Product Title Description": "Light Support, D/DD Cups",
+      "New or Used": "",
+      price: "$74 CAD",
+      "Review Score": 4.6,
+    },
+    {
+      "Origin URL":
+        "https://shop.lululemon.com/p/skirts-and-dresses-skirts/Pace-Rival-Skirt-Tall/_/prod3770001?color=60997",
+      "Product Features Limit": 4,
+      "Product Feature Sub Categories Limit": 4,
+      "Reviews of Product Limit": 5,
+      "Gender Cloths": "Women's Clothes",
+      "Type of Clothes": "Skirts",
+      "Product Title": "Pace Rival Mid-Rise Skirt",
+      "Extra Product Title Description": "Long",
+      "New or Used": "NEW",
+      price: "$78 CAD",
+      "Review Score": 4.4,
+    },
+    {
+      "Origin URL":
+        "https://shop.lululemon.com/p/women-tanks/Sleeveless-Polo-Shirt/_/prod10930138?color=60997",
+      "Product Features Limit": 4,
+      "Product Feature Sub Categories Limit": 4,
+      "Reviews of Product Limit": 5,
+      "Gender Cloths": "Women's Clothes",
+      "Type of Clothes": "Shirts",
+      "Product Title": "Quick-Dry Sleeveless Polo Shirt",
+      "Extra Product Title Description": "",
+      "New or Used": "",
+      price: "$68 CAD",
+      "Review Score": 4,
+    },
+    {
+      "Origin URL":
+        "https://shop.lululemon.com/p/jackets-and-hoodies-jackets/Hooded-Define-Jacket-Nulu-2/_/prod11500123?color=61860",
+      "Product Features Limit": 4,
+      "Product Feature Sub Categories Limit": 4,
+      "Reviews of Product Limit": 5,
+      "Gender Cloths": "Women's Clothes",
+      "Type of Clothes": "Coats & Jackets",
+      "Product Title": "Hooded Define Jacket",
+      "Extra Product Title Description": "Nulu",
+      "New or Used": "NEW",
+      price: "$138 CAD",
+      "Review Score": 4.8,
+    },
+    {
+      "Origin URL":
+        "https://shop.lululemon.com/p/women-sports-bras/Flow-Y-Bra-Nulu-Long-Line/_/prod9270365?color=56852",
+      "Product Features Limit": 4,
+      "Product Feature Sub Categories Limit": 4,
+      "Reviews of Product Limit": 5,
+      "Gender Cloths": "Women's Clothes",
+      "Type of Clothes": "Sports Bras",
+      "Product Title": "Flow Y Nulu Longline Bra",
+      "Extra Product Title Description": "Light Support, A–C Cups",
+      "New or Used": "",
+      price: "$64 CAD",
+      "Review Score": 3.7,
+    },
+    {
+      "Origin URL":
+        "https://shop.lululemon.com/p/womens-sweatpants/Studio-Crop-III-Unlined/_/prod2040016?color=56496",
+      "Product Features Limit": 4,
+      "Product Feature Sub Categories Limit": 4,
+      "Reviews of Product Limit": 5,
+      "Gender Cloths": "Women's Clothes",
+      "Type of Clothes": "Pants",
+      "Product Title": "Dance Studio Mid-Rise Cropped Pant",
+      "Extra Product Title Description": "",
+      "New or Used": "",
+      price: "$88 CAD",
+      "Review Score": 4.3,
+    },
+  ];
+  return JSON.stringify(products);
+  // const data = csvParse(productCatalog);
+  // const result = data.map(
+  //   (d) =>
+  //     `name: ${d["Product Title"]}\nurl: ${d["Origin URL"]}\ngender: ${d["Gender Cloths"]}\nprice: ${d["price"]}\nimage: ${d["Image URL"]}]}`
+  // );
+  // console.log(data.columns);
+  // console.log(result);
+  // return result;
+}
