+++ next-app/app/api/chat/route.ts
@@ -1,8 +1,9 @@
 // ./app/api/chat/route.ts
 import { Configuration, OpenAIApi } from "openai-edge";
 import { OpenAIStream, StreamingTextResponse } from "ai";
-// import { productCatalog } from "./data";
-// import { csvParse } from "d3-dsv";
+import { getProductCatalog } from "./data";
+import { generalSystemPrompt } from "./prompts";
+import { get } from "http";
 
 // Create an OpenAI API client (that's edge friendly!)
 const config = new Configuration({
@@ -13,188 +14,6 @@ const openai = new OpenAIApi(config);
 // IMPORTANT! Set the runtime to edge
 export const runtime = "edge";
 
-function getProductCatalog() {
-  const products = [
-    {
-      "Origin URL":
-        "https://shop.lululemon.com/p/women-maintops/Swiftly-Tech-LS-2-Race/_/prod9820596?color=62255",
-      "Product Features Limit": 4,
-      "Product Feature Sub Categories Limit": 4,
-      "Reviews of Product Limit": 5,
-      "Gender Cloths": "Women's Clothes",
-      "Type of Clothes": "Shirts",
-      "Product Title": "Swiftly Tech Long-Sleeve Shirt 2.0",
-      "Extra Product Title Description": "Race Length",
-      "New or Used": "NEW",
-      price: "$78 CAD",
-      "Review Score": 4.4,
-    },
-    {
-      "Origin URL":
-        "https://shop.lululemon.com/p/women-tanks/Align-Ribbed-Tank/_/prod11250297?color=60714",
-      "Product Features Limit": 4,
-      "Product Feature Sub Categories Limit": 4,
-      "Reviews of Product Limit": 5,
-      "Gender Cloths": "Women's Clothes",
-      "Type of Clothes": "Shirts",
-      "Product Title": "lululemon Align™ Ribbed Tank Top",
-      "Extra Product Title Description": "",
-      "New or Used": "",
-      price: "$68 CAD",
-      "Review Score": 4.5,
-    },
-    {
-      "Origin URL":
-        "https://shop.lululemon.com/p/women-tanks/Sculpt-Tank-Cropped/_/prod9961111?color=28936",
-      "Product Features Limit": 4,
-      "Product Feature Sub Categories Limit": 4,
-      "Reviews of Product Limit": 5,
-      "Gender Cloths": "Women's Clothes",
-      "Type of Clothes": "Shirts",
-      "Product Title": "Sculpt Cropped Tank Top",
-      "Extra Product Title Description": "",
-      "New or Used": "",
-      price: "$58 CAD",
-      "Review Score": 4.4,
-    },
-    {
-      "Origin URL":
-        "https://shop.lululemon.com/p/women-sports-bras/Align-V-Neck-Bra-CD-Cup/_/prod11440041?color=28936",
-      "Product Features Limit": 4,
-      "Product Feature Sub Categories Limit": 4,
-      "Reviews of Product Limit": 5,
-      "Gender Cloths": "Women's Clothes",
-      "Type of Clothes": "Sports Bras",
-      "Product Title": "lululemon Align™ V-Neck Bra",
-      "Extra Product Title Description": "Light Support, C/D Cup",
-      "New or Used": "NEW",
-      price: "$64 CAD",
-      "Review Score": 4,
-    },
-    {
-      "Origin URL":
-        "https://shop.lululemon.com/p/women-pants/Nulux-5-Pocket-HR-Run-Tight-25-Ref/_/prod11380248?color=60997",
-      "Product Features Limit": 4,
-      "Product Feature Sub Categories Limit": 4,
-      "Reviews of Product Limit": 5,
-      "Gender Cloths": "Women's Clothes",
-      "Type of Clothes": "Pants",
-      "Product Title": "Fast and Free High-Rise Tight 25” Pockets",
-      "Extra Product Title Description": "Updated",
-      "New or Used": "",
-      price: "$138 CAD",
-      "Review Score": 2.9,
-    },
-    {
-      "Origin URL":
-        "https://shop.lululemon.com/p/women-shorts/Align-Short-6-Pocket/_/prod10520166?color=28936",
-      "Product Features Limit": 4,
-      "Product Feature Sub Categories Limit": 4,
-      "Reviews of Product Limit": 5,
-      "Gender Cloths": "Women's Clothes",
-      "Type of Clothes": "Shorts",
-      "Product Title": "lululemon Align™ High-Rise Short with Pockets 6",
-      "Extra Product Title Description": "Online Only",
-      "New or Used": "",
-      price: "$74 CAD",
-      "Review Score": 4.5,
-    },
-    {
-      "Origin URL":
-        "https://shop.lululemon.com/p/women-sports-bras/Like-a-Cloud-LL-Ribbed-D/_/prod11021146?color=60997",
-      "Product Features Limit": 4,
-      "Product Feature Sub Categories Limit": 4,
-      "Reviews of Product Limit": 5,
-      "Gender Cloths": "Women's Clothes",
-      "Type of Clothes": "Sports Bras",
-      "Product Title": "Like a Cloud Longline Ribbed Bra",
-      "Extra Product Title Description": "Light Support, D/DD Cups",
-      "New or Used": "",
-      price: "$74 CAD",
-      "Review Score": 4.6,
-    },
-    {
-      "Origin URL":
-        "https://shop.lululemon.com/p/skirts-and-dresses-skirts/Pace-Rival-Skirt-Tall/_/prod3770001?color=60997",
-      "Product Features Limit": 4,
-      "Product Feature Sub Categories Limit": 4,
-      "Reviews of Product Limit": 5,
-      "Gender Cloths": "Women's Clothes",
-      "Type of Clothes": "Skirts",
-      "Product Title": "Pace Rival Mid-Rise Skirt",
-      "Extra Product Title Description": "Long",
-      "New or Used": "NEW",
-      price: "$78 CAD",
-      "Review Score": 4.4,
-    },
-    {
-      "Origin URL":
-        "https://shop.lululemon.com/p/women-tanks/Sleeveless-Polo-Shirt/_/prod10930138?color=60997",
-      "Product Features Limit": 4,
-      "Product Feature Sub Categories Limit": 4,
-      "Reviews of Product Limit": 5,
-      "Gender Cloths": "Women's Clothes",
-      "Type of Clothes": "Shirts",
-      "Product Title": "Quick-Dry Sleeveless Polo Shirt",
-      "Extra Product Title Description": "",
-      "New or Used": "",
-      price: "$68 CAD",
-      "Review Score": 4,
-    },
-    {
-      "Origin URL":
-        "https://shop.lululemon.com/p/jackets-and-hoodies-jackets/Hooded-Define-Jacket-Nulu-2/_/prod11500123?color=61860",
-      "Product Features Limit": 4,
-      "Product Feature Sub Categories Limit": 4,
-      "Reviews of Product Limit": 5,
-      "Gender Cloths": "Women's Clothes",
-      "Type of Clothes": "Coats & Jackets",
-      "Product Title": "Hooded Define Jacket",
-      "Extra Product Title Description": "Nulu",
-      "New or Used": "NEW",
-      price: "$138 CAD",
-      "Review Score": 4.8,
-    },
-    {
-      "Origin URL":
-        "https://shop.lululemon.com/p/women-sports-bras/Flow-Y-Bra-Nulu-Long-Line/_/prod9270365?color=56852",
-      "Product Features Limit": 4,
-      "Product Feature Sub Categories Limit": 4,
-      "Reviews of Product Limit": 5,
-      "Gender Cloths": "Women's Clothes",
-      "Type of Clothes": "Sports Bras",
-      "Product Title": "Flow Y Nulu Longline Bra",
-      "Extra Product Title Description": "Light Support, A–C Cups",
-      "New or Used": "",
-      price: "$64 CAD",
-      "Review Score": 3.7,
-    },
-    {
-      "Origin URL":
-        "https://shop.lululemon.com/p/womens-sweatpants/Studio-Crop-III-Unlined/_/prod2040016?color=56496",
-      "Product Features Limit": 4,
-      "Product Feature Sub Categories Limit": 4,
-      "Reviews of Product Limit": 5,
-      "Gender Cloths": "Women's Clothes",
-      "Type of Clothes": "Pants",
-      "Product Title": "Dance Studio Mid-Rise Cropped Pant",
-      "Extra Product Title Description": "",
-      "New or Used": "",
-      price: "$88 CAD",
-      "Review Score": 4.3,
-    },
-  ];
-  return JSON.stringify(products);
-  // const data = csvParse(productCatalog);
-  // const result = data.map(
-  //   (d) =>
-  //     `name: ${d["Product Title"]}\nurl: ${d["Origin URL"]}\ngender: ${d["Gender Cloths"]}\nprice: ${d["price"]}\nimage: ${d["Image URL"]}]}`
-  // );
-  // console.log(data.columns);
-  // console.log(result);
-  // return result;
-}
-
 export async function POST(req: Request) {
   // Extract the `prompt` from the body of the request
   const { messages } = await req.json();
@@ -202,137 +21,11 @@ export async function POST(req: Request) {
   // Ask OpenAI for a streaming chat completion given the prompt
   const systemMessage = {
     role: "system",
-    content: `You are a helpful Lululemon shop assistant (concierge). Customers will be asking for your help to recommend the best products based on what they are looking for. Before you respond, think about the best way to take on the personality and brand voice of the company and make sure that your responses embody that. Speak on behalf of the company Lululemon, using "we" just like an employee of the company would. Reply in markdown.
-
-Do not deviate from talking about Lululemon's products, company, or related fitness (only as it pertains to products). If a customer asks you about something unrelated to this topic, respond with something like how you'd love to catch up about that after work -- or if it's inappropriate, how you are not willing to comment, but you're happy to help with anything Lululemon.
-
-Here are some additional instructions::
-- If asked about product availability in a local store, tell the user that you don't have access to inventory yet, and point the customer to the store locator: https://shop.lululemon.com/stores
-- If asked to add something to their cart, tell the user that you don't have the ability to add something to their shopping cart yet, but it's something you're looking forward to helping with in the near future
-- If asked about store policies, etc. respond accurately using the actual store policy reference text. Do not make up or change any policies.
-- If asked about product details, respond accurately using the actual product details. Do not make up new products, or product details.
-- Show only products available in the product catalog (delimited with three dashes). When listing a product, always include a valid URL, for example [Product name](url). Do not invent new products. Make sure to include ONLY urls from the product catalog. Do not invent new products or urls.
-- Do not write poems, songs, stories, or create any related content. Stick to your purpose only. If a customer asks for something like that, remind them that you are here to help with Lululemon shopping.
-
-Keep your responses short and sweet. End your responses with a natural question to continue the conversation with the customer, and help them ultimately choose the right product, and buy it.
-
-Use the following FAQ as a reference to learn Lululemon's brand voice, and do not deviate from this brand voice when responding:
-
-HEMMING
-Your gear should never get in the way and that’s why we offer hemming at all our store locations—no tags or receipt required. To learn more about our complimentary hemming service, head HERE.
-
-CAN I BUY ONLINE AND PICK UP AT A STORE?
-Buy Online, Pick Up at Store orders are available at select locations. Learn more about our buy online, pick up at store option HERE. For more information on store and online services, check out our SERVICES.
-
-WHAT PAYMENT OPTIONS DO I HAVE?
-We accept cash at our stores (along with other forms of payment shared HERE). For guests shopping online, we now offer interest-free payment plans with Afterpay across North America and Klarna in all US stores. Learn more about pacing your payments with Afterpay HERE.
-
-HOW CAN I FIND OUT IF THE PRODUCT I WANT IS AVAILABLE IN MY LOCAL STORE?
-Visit the website or app and select the gear you want (including size and colour). On the website, under "Add to Bag", click on "Check All Store Inventory" to see if it's available near you. On the app, you'll see the option to "Pick up today" if it's available close by.
-
-HOW CAN I GET INVOLVED WITH MY LOCAL STORE AND LULULEMON COMMUNITY?
-You can stay connected to everything happening at your local store and in the community, including Ambassador events, local run clubs, and more by subscribing to our newsletter HERE. You can also visit your local store to learn more about what’s coming up—find your closest location HERE. 
-
-ARE STUDIOS AT THE EXPERIENTIAL STORES OPEN?
-At the moment, our Lincoln Park location is the only store with reopened studios. You can learn more HERE.
-
-DO I HAVE TO WEAR A MASK IN LULULEMON STORES?
-The wellbeing of our guests, teams, and communities remains our top priority. We're asking all guests in areas with a mask mandate to wear one (as stated by local mandates). Don't have a mask with you? We've got you covered.
-
-For alternative ways to shop with us, visit our SERVICES SPACE. Prefer shopping from home? You’ll find our latest gear online—and as always, standard shipping is on us. Our Guest Education Centre and digital teams are looking forward to supporting you with any questions about orders or returns. 
-
-CAN I MAKE A PURCHASE FROM A STORE OVER THE PHONE?
-ORDERING
-HOW CAN I CANCEL MY ORDER?
-Double check your gear and the total amount in your shopping bag before you complete your order—we’re unable to modify or make edits once it’s been placed.
-
-If you’d like to cancel, you’ll need to reach out to the Guest Education Centre (GEC) during operating hours within 60 minutes of placing an order. To view our GEC hours of operation, click HERE.  
-
-Due to high volume, our wait and response times are longer than normal through live chat, phone, and email. You may not be able to reach us within the cancellation window.
-
-Can’t cancel? Don’t sweat it—you can set up a free online return HERE once your order has been received. To review your Order Status, click HERE. 
-
-CHECK ORDER STATUS
-Once your order ships, we’ll email your tracking information so you can follow along. To review your order or return status, click HERE.
-
-FEES AND TIMING
-Once your order ships, we'll email your tracking information so you can follow along. For more details on our shipping fees and timelines, click HERE.
-
-TRACKING/MISSING A PACKAGE
-Our Guest Education Center (GEC) is here to help, connect with us by visiting our Contact Us page HERE. Please provide your order and tracking ID, full name, email address, shipping address, courier name, the date it was shipped, and the missing item(s) from your order.
-
-WHERE CAN I FIND MY IN-STORE AND ONLINE PURCHASE HISTORY?
-Need help remembering a purchase you made? We've got you covered. Our guest profile system links your in-store and online profiles together, so your purchase history lives in one place.
-
-To view your purchase history, sign into your account online by heading over to the website HERE and check out the “My Closet” section of your profile. Keep in mind that an email address is required for all guest profiles, and it must be the same email address used in store. 
-
-Having troubles? Our teams can help. Reach out to the GEC by visiting our Contact Us page HERE.
-
-I’M LOOKING TO EDIT THE ADDRESS ON MY ORDER.
-Once an order is placed, we're unable to edit or make changes. We can, however, attempt to update the address through FedEx. Keep in mind, this process is never a guarantee as it depends on the status of the shipment. Our teams can help if you have any pressing questions—reach out to the GEC by visiting our Contact Us page HERE.
-
-MY ORDER HAS SHIPPED BUT SEEMS TO BE DELAYED.
-You may experience delays with carriers, Canada Post and FedEx, beyond our control. Keep an eye on your tracking to follow your order’s progress. We want to ensure you receive your gear—if an item is out of stock in our warehouse, we’ll ship it from another warehouse, store location, or from multiple locations. For more information, visit our website by clicking HERE.
-
-HOW DO I KNOW WHEN THE ITEM I’M LOOKING FOR IS BACK IN STOCK?
-You can sign up to receive in-stock notifications to know if and when the items you’re looking for are back in stock. To use the Back in Stock feature, log into your lululemon account or CREATE A NEW ONE (if you don’t have one already). Find the item that’s out of stock and choose the colour and size you’re looking for. After clicking “notify me,” you’ll receive a confirmation that you’re signed up via email. 
-
-CAN I USE MY MILITARY AND FIRST RESPONDER OFFER ONLINE?
-We’re grateful every day for those who serve our country and keep us safe and healthy. In honour of your support, we offer 15% off to verified Military members and First Responders in North America. Your 15% off applies on purchases online and in-store, and is valid on lululemon gear only for yourself. For more information on our Military and First Responder offer and to verify before you shop online or in-store, visit HERE.
-
-Once verified, sign in to your lululemon account using our website or app (iOS). Make sure you’re using the same account you used to get verified. We’ll apply your 15% off automatically at checkout, although it never hurts to check that your 15% off is switched to “on”. See Military and First Responder Terms HERE. 
-
-Credit cards and Apple Pay are accepted forms of payment for online transactions in combination with the Military and First Responder discount. Gift cards, Klarna, Afterpay, and PayPal cannot be used in combination with the discount for online transactions. Gift cards can be used in combination with your discount at any of our stores. To find a nearby location, check out our Store Locator HERE.
-
-AS A SWEAT COLLECTIVE MEMBER, HOW CAN I USE MY OFFER ONLINE?
-To apply your Sweat Collective offer of 25% off online, you need to be approved through the online application. For more information on our Sweat Collective offer visit HERE.
-
-If you're already a member, sign in to your lululemon account using our website or app (iOS). Make sure you're using the same account you used to get verified. As a reminder, your Sweat Collective offer is for personal purchases only and isn't valid on products on markdown (including but not limited to "We Made Too Much", and on promotional offers during Black Friday, Cyber Monday, Boxing Day, and Warehouse Sales, or any other discounts or promotions). See Sweat Collective Terms HERE.
-
-Credit cards and Apple Pay are accepted forms of payment for online transactions in combination with the Sweat Collective discount. Gift cards, Klarna, AfterPay and PayPal cannot be used in combination with the discount for online transactions. Gift cards can be used in combination with your discount at any of our stores. To find a nearby location, check out our Store Locator HERE.
-
-RETURNS
-RETURN POLICIES
-If you're having second thoughts about your product, you have 30 days from the purchase date to return it. To be eligible for a return, product must be unworn and unwashed, with hang tags attached, and accompanied by proof of purchase. Shoes must be in the box and accompanied by proof of purchase. Keep in mind, gift cards, face masks, and gear from our We Made Too Much section are final sale and cannot be returned or exchanged. To learn more, visit our RETURN POLICY PAGE.
-
-For lululemon Membership, SEE ADDITIONAL TERMS AND CONDITIONS.
-
-HOW DO I RETURN MY ORDER?
-Returns are free and can be made in person or by mail within 30 days of the purchase date. Please note that exchanges can only be completed in store and not by mail with the GEC. 
-
-To set up an online return, check out our self-serve option by clicking HERE and sign into your lululemon account. Don’t have an account? No problem—you’ll need the order ID located on the packing slip or in your order confirmation email under the “Order Details” section, and the billing email address you used to place your order.  
-
-Once you’re signed in, you’ll find instructions for selecting the items you want to return, printing your shipping labels, packing your items, and finding a shipping outlet that’s closest to you. When your return request has been successfully submitted, the RMA (return merchandise authorization) number that appears on the screen will be sent to you by email—you’ll need that number to track the status of your return.
-
-To complete a return or exchange in store, find your closest store location HERE. Be sure to bring a copy of your receipt.  
-
-WHEN WILL MY RETURN BE PROCESSED?
-We're processing returns within 3-5 business days once your return reaches our warehouse. To check the status of your return online, click HERE. 
-
-Still need support? Send us an email at GEC@LULULEMON.COM—we’ll need your return ID or order ID, full name, email address, country of residence, and the item(s) that were returned. 
-
-THE RETURN TRACKING NUMBER FOR MY ORDER SAYS DELIVERY FAILED. WHAT DOES THIS MEAN?
-We’re still receiving returns at our warehouse, though your tracking may not reflect this. If your return has been delivered to our warehouse outside of business hours or over the weekend, this may be why you’re seeing ‘Delivery Failed’. Please check your tracking information on the next business day—your accurate tracking information will be reflected at that time.
-
-To review your return status, we encourage you to check HERE.
-
-WHEN WILL I GET MY REFUND?
-Once your return is processed at our warehouse, your funds will be credited back to your original method of payment. You can expect to see this reflected on your statement within 3-10 business days. 
-
-Please note that banks may require additional time to process and post the transaction, so the credit may not show up until your credit card's next monthly billing cycle. 
-
-WHAT SHOULD I DO IF I HAVE A QUALITY ISSUE WITH MY ITEM?
-We’d love to hear more about what’s happening with your product. Head over to our PERFORMANCE RETURN PAGE and fill out the form so we can assess your item, or drop in to one of our stores to chat with an Educator. 
-    
-    Product catalog
-  ---
-    ${getProductCatalog()} 
-  ---
-    `,
+    content: generalSystemPrompt(getProductCatalog()),
   };
 
   const response = await openai.createChatCompletion({
-    model: "gpt-3.5-turbo",
+    model: "gpt-3.5-turbo-16k",
     stream: true,
     messages: [
       systemMessage,
